+++ src/main/java/dta/sfmflow/client/screen/widgets/AdvancedItemFilterVariableSettingsOverlay.java
@@ -57,7 +57,7 @@ public class AdvancedItemFilterVariableSettingsOverlay extends NodeSettingsOverl
 		this.width = 240;
 		this.height = 195;
 		this.setX((parentScreen.width - this.width) / 2);
-		this.setY(parentScreen.getTopPos() + (256 - this.height) / 2);
+		this.setY(parentScreen.getOverlayTargetY(this.height));
 
 		parentScreen.getMenu().setActiveComponent(component);
 		PacketDistributor.sendToServer(new SetActiveFilterComponentPacket(