+++ src/main/java/dta/sfmflow/client/screen/widgets/FluidTransferSettingsOverlay.java
@@ -32,7 +32,7 @@ public class FluidTransferSettingsOverlay extends NodeSettingsOverlay {
 		this.width = 300;
 		this.height = 360;
 		this.setX((parentScreen.width - 300) / 2);
-		this.setY(25);
+		this.setY(parentScreen.getOverlayTargetY(this.height));
 
 		component.setUseAll(false);
 		component.setTargetSlot(-1);